refactor(desktop): extract picker controller + linked-context hooks from PromptGroup
PromptGroup was 592 lines mixing form state, agent prefs, branch picker
state + handlers, link handlers, hotkey wiring, and JSX. Split the
picker controller and link handlers into co-located hooks; PromptGroup
is now 360 lines focused on form composition.

- useBranchPickerController: owns picker state (search, filter), the
  branch-context query, host-id resolution + workspace lookup, and the
  three per-row action handlers (Open / Check out / Adopt). Returns a
  `pickerProps` bag the caller spreads into <CompareBaseBranchPicker>.
  Hides ~70 lines of plumbing behind one hook call.
- useLinkedContext: bundles addLinkedIssue / addLinkedGitHubIssue /
  removeLinkedIssue / setLinkedPR / removeLinkedPR. Pure delegation
  over updateDraft.
- Collapse the PromptGroup / PromptGroupInner indirection. There was
  no error boundary, provider, or conditional in the wrapper — just a
  pass-through. Removed.

Behavior unchanged. Typecheck + lint clean; 11 pending-page tests pass.
