+++ apps/desktop/src/renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/PromptGroup/hooks/useBranchPickerController/index.ts
@@ -1,0 +1,1 @@
+export { useBranchPickerController } from "./useBranchPickerController";
